src/CustomerChatbotAPI/pyproject.toml
[project]
name = "customer-chatbot-api"
version = "0.1.0"
description = "Customer Chatbot GSA with Voice — multi-agent conversational AI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sas-cosmosdb>=0.1.4",
    "sas-storage>=0.1.0",
    "semantic-kernel>=1.15.0",
    "azure-ai-projects>=1.0.0b5",
    "azure-search-documents>=11.6.0",
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.9.0",
    "openai>=1.55.0",
    "websockets>=13.0",
    "pyjwt[crypto]>=2.9.0",
    "httpx>=0.27.0",
    "azure-cognitiveservices-speech>=1.40.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.2.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = ["--import-mode=importlib"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
